+++ types/fetch-node.d.ts
@@ -1,7 +1,7 @@
 export class FetchNodeSession extends FetchSession {
     constructor(apiUrl: any, defaultParams?: {}, defaultOptions?: {});
     agent: CookieAgent;
-    getFetchOptions(headers: any): any;
+    getFetchOptions(fetchOptions: any): any;
 }
 import { FetchSession } from './fetch.js';
 import { CookieAgent } from 'http-cookie-agent/undici';